ERRORS- Large directory changes
our app router was not working as it should because our directories are not in the correct locations. Due to this i cant implement proper session provider into the layout file. Had to change a lot of things, right now the website loads up to the login page. DO NOT MERGE THIS INTO ANYTHING.